+++ include/dlaf/gpu/blas/error.h
@@ -18,7 +18,7 @@
 #include "dlaf/common/source_location.h"
 #include "dlaf/gpu/blas/api.h"
 #include "dlaf/gpu/cublas/error.h"
-#include "dlaf/gpu/hipblas/error.h"
+#include "dlaf/gpu/rocblas/error.h"
 
 namespace dlaf::gpublas::internal {
